プロフィールのコンテンツをコンポーネント化
